Importance of NetworkTransformChild
With my complicated Player Prefab, i found you need a network transform child for each gameobject's child that you want to be updated through the server to all the other clients, otherwise their positions do not move with the parent to other clients.
Very cool discovery and will have more fun with this mini-game later on.